Task 1.2, alternative implementation
