+++ src/main/scala/com/gd/AttributionNodeAgg.scala
@@ -1,0 +1,6 @@
+package com.gd
+import org.apache.spark.sql.DataFrame
+
+class AttributionNodeAgg(cfg: AttributionNodeConfiguration) extends JoinNode {
+  override def join(left: DataFrame, right: DataFrame): DataFrame = ???
+}
